// src/client/interfaces/Commande.ts
interface Produit {
    thumbnail:string;
    nom_produit:string;
}

interface DetailCommande {
    produit : Produit;
    quantite: string;
}

export interface Commande {
    identifiant_commande: string;
    code_livraison: string;
    date_commande: string;
    etat_commande: string;
    lieu_livraison: string;
    total_ht: string;
    total_ttc: string;
    frais_livraison_appliques?: string;
    details_commandes?: DetailCommande[];
}